refactor: clean up router.ts
